eval/integration/requirements-integration.txt
# Integration example dependencies (not part of core library)
fastapi>=0.110
uvicorn[standard]>=0.27
requests>=2.31
python-dotenv>=1.0
